+++ src/examples/utils/model.h
@@ -8,10 +8,10 @@ typedef struct Model
     float*    vertices;
     uint32_t  vertex_count;
     uint8_t   is_indexed;
-    uint32_t  vertex_buffer_offset;
+    uint64_t  vertex_buffer_offset;
     uint16_t* indices;
     uint32_t  index_count;
-    uint32_t  index_buffer_offset;
+    uint64_t  index_buffer_offset;
 } Model;
 
 void clear_model(Model * model);
